settings: refactor handling of storages with meta-data/tombstones
Currently, meta-data has a very narrow use: as tombstones.

Later, we will need to store additional per UUID meta-data. For example,
when a profile becomes unsaved, we may need to remember the original
filename.

Refactor the code for that. This is for the most part just renaming
and slightly different handling of the fields.