device/wifi: various cleanup in act_stage1_prepare()
The only change in behavior is in act_stage1_prepare().

That changes compared to before that we also set the specific
object path if it was already set (and we looked up the AP by
specific object to start with).

Also, for existing APs that we found with nm_wifi_aps_find_first_compatible(),
it changes the order of calling set_current_ap() before nm_active_connection_set_specific_object().

That should not make a different though. I anyway wonder why we even bother to
set the specific object on the AC. Maybe that should be revisited.